imdb_project/requirements.txt
pyspark==3.2.1
boto3
requests
pandas
apache-airflow
SQLAlchemy
sqlalchemy-utils
python-dateutil
python-dotenv